modularize create app file
