+++ test/colorset.test.js
@@ -1,0 +1,7 @@
+const assert = require("assert");
+const colorSet = require("../lib/colorsets");
+
+/**
+ * will write and set tests later on
+ */
+assert.strictEqual(Object.length(colorSet), 3)
